requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
groq==0.4.2
asyncpg==0.29.0
sqlalchemy==2.0.25
redis==5.0.1
influxdb-client==1.39.0
paho-mqtt==2.0.0
python-dotenv==1.0.0
sse-starlette==1.8.2
python-json-logger==2.0.7
slowapi==0.1.9
pytest==7.4.4
pytest-asyncio==0.23.3
psutil==5.9.8
# ML dependencies for anomaly detection & device classification
scikit-learn==1.4.0
numpy==1.26.3
xgboost==2.0.3
